+++ query/Cargo.toml
@@ -54,6 +54,7 @@ common-tracing = { path = "../common/tracing" }
 # databend-query dependencies
 query-common = { path = "common" }
 query-configs = { path = "configs" }
+query-flight = { path = "flight" }
 query-users = { path = "users" }
 
 # Github dependencies
